requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
redis==4.0.2
flask-login==0.5.0
requests==2.26.0
pytest==7.0.1
PyJWT==2.3.0
stripe==2.65.0